feat(#440): migrate BulkTagModal to react-hook-form + Zod + TypeScript
